backend/requirements.txt
fastapi==0.115.12
numpy==2.2.4
openai==1.70.0
pandas==2.2.3
pandasai==2.4.2
pandasai_openai==0.1.5
python-dotenv==1.1.0
uvicorn==0.34.0
